fixes some minor bugs on poster details page, and update some ui.

+++ Backend/src/stripe-webhook/stripe-webhook.service.ts
@@ -1,6 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { log } from 'console';
+import { BookingService } from 'src/booking/booking.service';
 import { CartService } from 'src/cart/cart.service';
+import { UpdatePosterDto } from 'src/poster-details/dto/updatePoster.dto';
+// import { UpdatePosterDto } from 'src/poster-details/dto/updatePoster.dto';
+import { PosterDetailsService } from 'src/poster-details/poster-details.service';
 import { Stripe } from 'stripe';
 
 @Injectable()
@@ -10,6 +15,8 @@ export class StripeWebhookService {
   constructor(
     private configService: ConfigService,
     private cartService: CartService,
+    private bookingService: BookingService,
+    private posterService: PosterDetailsService,
   ) {
     this.stripe = new Stripe(this.configService.get('STRIPE_SECRET_KEY'), {
       apiVersion: '2024-04-10', // Specify the Stripe API version
@@ -36,41 +43,61 @@ export class StripeWebhookService {
         // Handle successful payment
         const session = event?.data?.object;
         const { metadata, mode, payment_intent, invoice } = session;
-        console.log(
-          'payment_intent ========================',
-          metadata.purchased_items,
-        );
         const cartId = JSON.parse(metadata.purchased_items);
         if (cartId && cartId.length > 0) {
           try {
             for (const itemId of cartId) {
-              // Delete cart item based on purchased item ID
-              await this.cartService.deleteCartById(itemId);
+              // get cart data
+              const cartItem = await this.cartService.getCartById(itemId);
+              if (cartItem) {
+                const filteredCartItem = {
+                  userId: cartItem.userId,
+                  posterId: cartItem.posterId,
+                  title: cartItem.title,
+                  image: cartItem.image,
+                  address: cartItem.address,
+                  totalPrice: cartItem.totalPrice,
+                  bookingDate: cartItem.bookingDate,
+                };
+
+                const booking =
+                  await this.bookingService.createBooking(filteredCartItem);
+                if (!booking) {
+                  console.log('booking failed----------', booking);
+                  return;
+                }
+                console.log('Booking created:', booking);
+
+                // updating the booking dates in the posters-----------------------
+                const updatePosterDto: UpdatePosterDto = {
+                  bookingDate: cartItem.bookingDate, // Example array of booking dates
+                };
+                this.posterService.updatePoster(
+                  cartItem.posterId,
+                  updatePosterDto,
+                );
+                // Delete cart data-----------------------------
+                const deleted = await this.cartService.deleteCartById(itemId);
+                if (!deleted) {
+                  console.log(
+                    'cart items is not deleted or their is no items in the cart to delete.',
+                  );
+                  return;
+                }
+                console.log('Cart item deleted successfully:', cartId);
+              } else {
+                console.log('Cart item not found for ID:', cartId);
+              }
             }
-            console.log('Cart items deleted successfully.');
           } catch (error) {
-            console.error('Error deleting cart items:', error);
+            console.error('Error deleting cart items & booking orders:', error);
           }
         } else {
           console.log('No purchased item IDs found in metadata.');
         }
-
-        // await this.handleSuccessfulPayment(event.data.object);
         break;
-      // Add other event types as needed
       default:
-        // Handle other event types or log unrecognized events
         console.log('Unhandled event type:', event.type);
     }
   }
-  //   async handleSuccessfulPayment(paymentIntent: any) {
-  //     // Implement logic to handle successful payment
-  //     const userId = paymentIntent.metadata.userId;
-  //     const dataId = paymentIntent.metadata.dataId;
-  //     const totalPrice = paymentIntent.amount / 100; // Convert from cents to dollars
-
-  //     // Example: Call a service to delete cart data and create booking schema
-  //     // await this.cartService.deleteCart(userId);
-  //     // await this.bookingService.createBooking(userId, dataId, totalPrice);
-  //   }
 }
